% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_across.R
\name{multiply_across}
\alias{multiply_across}
\title{Multiplies percents across individual columns}
\usage{
multiply_across(data = data_select_weight, list = percent_to_1)
}
\arguments{
\item{data}{A data set}

\item{list}{A list of percentages to multiply across columns ofthe dataset}
}
\value{
A binded object of modified tabled that had multiplied across percentages
}
\description{
This function allows you to multiply individual columns by a list of percentages.
}
\examples{
multiply_across(df, list)
multiply_across(df, c(.05,.1))
multiply_across(data, list)
}
